[mojo-lang] Fix meta-trait type parameter rebinding logic. (#36584)
This enhances type checking of type parameter expressions of
AnyTrait type to look through downcasts of the type itself and
allows downcast from a value of `AnyTrait[SomeTrait]` to
`SomeTrait`.  This is enough to enable us to type check the
fancy dependently typed `each` method on `VariadicPack` which
makes VariadicPack useful for non-AnyType trait requirements
as shown in the integration test!!!

This bumps up against a bug in the KGEN ParameterVerifier.  I
believe the fix is simple, but am not 100% confident about it,
so I temporarily disabled it and filed #36583 to discuss this.

modular-orig-commit: 816c623595bce4a995afc795bf5fd551c0c83900