[mojo-lang] Reject `AnyRegType` variadic packs, pushing to `AnyType` (#36906)
`AnyRegType` variadic packs are a legacy feature that predated traits
(and possibly memory only types in general).  We are unifying on a
single consistent representation built around `VariadicPack`.

This takes the step of banning the old `AnyRegType` form in the
parser.  That said, we still have a few uses of the old form in the
tree, so it keeps them alive when explicitly declared "borrowed".
This will allow us to work through removing/replacing these cases
incrementally.

modular-orig-commit: 8e7642b4396f704020adabf48c2e7dfb01ec6e9d